Storing ray generation in IntersectionContext, and added RayGeneration texture to use it; Shader and RayTracer::estimateLightContribution: adding a subtractiveHack flag for an illumination hack we're trying.
git-svn-id: https://svn.code.sf.net/p/liar/code/trunk/liar@173 b153b6fe-5711-0410-bcb9-8978480c73f3

+++ src/tracers/adjoint_photon_tracer.cpp
@@ -58,7 +58,7 @@ void AdjointPhotonTracer::doPreProcess(const kernel::TSamplerPtr&, const TimePer
 
 const XYZ AdjointPhotonTracer::doCastRay(
 		const kernel::Sample& sample, const kernel::DifferentialRay& primaryRay,
-		TScalar& tIntersection, TScalar& alpha, int generation, bool highQuality) const
+		TScalar& tIntersection, TScalar& alpha, size_t generation, bool highQuality) const
 {
 /*	TScalar weight = 1;
